refactor: colocate component-owned styles and constants
