+++ FateConnect/Web/src/pages/Rides/components/RideCard/RideDeleteConfirmation/constants/index.ts
@@ -1,0 +1,7 @@
+export const DELETE_DIALOG = {
+  title: 'Confirmar Exclusão',
+  messagePrefix: 'Tem certeza que deseja excluir a carona para ',
+  messageSuffix: '?',
+  confirmLabel: 'Excluir',
+  cancelLabel: 'Cancelar',
+};
