+++ FateConnect/Web/src/pages/Rides/constants/index.ts
@@ -14,24 +14,7 @@ export const RIDE_LIST_MESSAGES = {
   deleteFailed: 'Erro ao excluir a carona. Tente novamente.',
 };
 
-export const DELETE_DIALOG = {
-  title: 'Confirmar Exclusão',
-  messagePrefix: 'Tem certeza que deseja excluir a carona para ',
-  messageSuffix: '?',
-  confirmLabel: 'Excluir',
-  cancelLabel: 'Cancelar',
-};
-
 export const RIDE_CARD_LABELS = { edit: 'Editar', delete: 'Excluir', contact: 'Contato' };
-
-export const CONTACT_DIALOG = {
-  title: 'Informações de Contato',
-  emailCopied: 'E-mail copiado!',
-  emailCopyFailed: 'Não foi possível copiar o e-mail.',
-  /** Texto que já vai escrito na conversa, para quem recebe saber de qual carona se trata. */
-  message: (destination: string): string =>
-    `Olá! Vi no FateConnect a sua carona para ${destination} e queria saber se ainda tem vaga.`,
-};
 
 const SINGLE_SEAT = 1;
