+++ FateConnect/Web/src/pages/Signup/Signup.test.tsx
@@ -8,6 +8,9 @@ import { LandingSectionEnum, RoutePathEnum } from '@app/routes/paths';
 import { render, screen, userEvent, waitFor, within } from '@app/test/testing-library';
 
 import { Signup } from '.';
+import { PASSWORD_TOGGLE_LABEL } from './components/AccountSection/constants';
+import { CALENDAR_TOGGLE_LABEL } from './components/BirthDateField/constants';
+import { LEGAL_SOON_MESSAGES } from './components/ConsentSection/constants';
 import * as C from './constants';
 import { SIGNUP_MESSAGES } from './schema';
 
@@ -172,7 +175,7 @@ describe('Signup', () => {
   it('should toggle the password visibility and show the current state in the icon', async () => {
     renderSignup();
     const password = screen.getByLabelText(/^Senha/);
-    const toggle = screen.getByRole('button', { name: C.PASSWORD_TOGGLE_LABEL });
+    const toggle = screen.getByRole('button', { name: PASSWORD_TOGGLE_LABEL });
 
     expect(password).toHaveAttribute('type', 'password');
     expect(toggle).toHaveAttribute('aria-pressed', 'false');
@@ -231,11 +234,11 @@ describe('Signup', () => {
 
     await userEvent.click(screen.getByRole('button', { name: 'Termos de Uso' }));
 
-    expect(await screen.findByText(C.LEGAL_SOON_MESSAGES.terms)).toBeInTheDocument();
+    expect(await screen.findByText(LEGAL_SOON_MESSAGES.terms)).toBeInTheDocument();
 
     await userEvent.click(screen.getByRole('button', { name: 'Política de Privacidade' }));
 
-    expect(await screen.findByText(C.LEGAL_SOON_MESSAGES.privacy)).toBeInTheDocument();
+    expect(await screen.findByText(LEGAL_SOON_MESSAGES.privacy)).toBeInTheDocument();
   });
 
   it('should not toggle the consent when the legal link is clicked', async () => {
@@ -357,7 +360,7 @@ describe('Signup', () => {
     const birthDate = screen.getByLabelText(/Data de nascimento/);
     await userEvent.type(birthDate, '22051999');
 
-    await userEvent.click(screen.getByRole('button', { name: C.CALENDAR_TOGGLE_LABEL }));
+    await userEvent.click(screen.getByRole('button', { name: CALENDAR_TOGGLE_LABEL }));
     const calendar = await screen.findByRole('grid');
     await userEvent.click(within(calendar).getByRole('gridcell', { name: '10' }));
