+++ FateConnect/Web/src/pages/Signup/components/BirthDateField/constants/index.ts
@@ -1,0 +1,1 @@
+export const CALENDAR_TOGGLE_LABEL = 'Abrir calendário';
