+++ FateConnect/Web/src/pages/Signup/components/ConsentSection/constants/index.ts
@@ -1,0 +1,11 @@
+export const CONSENT_TERMS_PREFIX = 'Eu concordo com os';
+export const CONSENT_TERMS_LINK = 'Termos de Uso';
+export const CONSENT_TERMS_SEPARATOR = 'e';
+export const CONSENT_PRIVACY_LINK = 'Política de Privacidade';
+export const CONSENT_MARKETING_LABEL =
+  'Eu concordo em receber e-mails e notificações do aplicativo';
+
+export const LEGAL_SOON_MESSAGES = {
+  terms: 'Termos de uso estarão disponíveis em breve.',
+  privacy: 'Política de privacidade estará disponível em breve.',
+};
