+++ FateConnect/Web/src/pages/Signup/components/ConsentSection/index.tsx
@@ -3,10 +3,10 @@ import type { MouseEvent } from 'react';
 import { useFormContext } from 'react-hook-form';
 
 import { useNotification } from '@app/hooks/useNotification';
-import * as C from '@app/pages/Signup/constants';
 import type { SignupFormValues } from '@app/pages/Signup/schema';
 import { Checkbox, FormControlLabel } from '@design-system';
 
+import * as C from './constants';
 import * as S from './styles';
 
 /** Aceites. Os documentos legais ainda não existem — o clique avisa isso. */
