+++ FateConnect/Web/src/pages/Signup/components/ContactSection/styles.ts
@@ -1,0 +1,9 @@
+import { Box, desktopMedia, styled } from '@design-system';
+import type { PolymorphicProps } from '@design-system';
+
+/** Metade da linha no desktop: 3 de 6 colunas. */
+export const HalfWidthCell = styled(Box)<PolymorphicProps>({
+  width: '100%',
+
+  [desktopMedia]: { gridColumn: 'span 3' },
+});
