+++ FateConnect/Web/src/pages/Signup/constants/index.ts
@@ -1,7 +1,3 @@
-import type { SelectOption } from '@design-system';
-
-import { GenderValueEnum } from '../@types';
-
 export const SIGNUP_TITLE = 'Crie sua Conta';
 export const ADDRESS_SECTION_TITLE = 'Endereço';
 export const CONTACT_SECTION_TITLE = 'Dados para contato';
@@ -31,25 +27,10 @@ export const FIELD_PLACEHOLDERS = {
 };
 
 export const SELECT_PLACEHOLDER = 'Selecione...';
-export const PASSWORD_TOGGLE_LABEL = 'Mostrar ou ocultar senha';
-export const CALENDAR_TOGGLE_LABEL = 'Abrir calendário';
-export const ZIP_LOOKUP_LABEL = 'Consultando CEP';
-
-export const CONSENT_TERMS_PREFIX = 'Eu concordo com os';
-export const CONSENT_TERMS_LINK = 'Termos de Uso';
-export const CONSENT_TERMS_SEPARATOR = 'e';
-export const CONSENT_PRIVACY_LINK = 'Política de Privacidade';
-export const CONSENT_MARKETING_LABEL =
-  'Eu concordo em receber e-mails e notificações do aplicativo';
 
 export const SUBMIT_LABEL = 'Cadastrar';
 export const LOGIN_PROMPT = 'Já tem uma conta?';
 export const LOGIN_LINK_LABEL = 'Faça login';
-
-export const LEGAL_SOON_MESSAGES = {
-  terms: 'Termos de uso estarão disponíveis em breve.',
-  privacy: 'Política de privacidade estará disponível em breve.',
-};
 
 export const SIGNUP_ERROR_MESSAGES = {
   emailTaken: 'Este e-mail já está em uso.',
@@ -64,51 +45,3 @@ export const ZIP_LOOKUP_MESSAGES = {
 
 export const signupSuccessMessage = (fullName: string): string =>
   `Conta criada com sucesso, ${fullName}!`;
-
-export const GENDER_OPTIONS: readonly SelectOption[] = [
-  { value: GenderValueEnum.MALE, label: 'Masculino' },
-  { value: GenderValueEnum.FEMALE, label: 'Feminino' },
-  { value: GenderValueEnum.OTHER, label: 'Prefiro não informar' },
-];
-
-/** Unidades federativas (sigla + nome em pt-BR). */
-export const BRAZILIAN_STATES: readonly SelectOption[] = [
-  { value: 'AC', label: 'Acre' },
-  { value: 'AL', label: 'Alagoas' },
-  { value: 'AP', label: 'Amapá' },
-  { value: 'AM', label: 'Amazonas' },
-  { value: 'BA', label: 'Bahia' },
-  { value: 'CE', label: 'Ceará' },
-  { value: 'DF', label: 'Distrito Federal' },
-  { value: 'ES', label: 'Espírito Santo' },
-  { value: 'GO', label: 'Goiás' },
-  { value: 'MA', label: 'Maranhão' },
-  { value: 'MT', label: 'Mato Grosso' },
-  { value: 'MS', label: 'Mato Grosso do Sul' },
-  { value: 'MG', label: 'Minas Gerais' },
-  { value: 'PA', label: 'Pará' },
-  { value: 'PB', label: 'Paraíba' },
-  { value: 'PR', label: 'Paraná' },
-  { value: 'PE', label: 'Pernambuco' },
-  { value: 'PI', label: 'Piauí' },
-  { value: 'RJ', label: 'Rio de Janeiro' },
-  { value: 'RN', label: 'Rio Grande do Norte' },
-  { value: 'RS', label: 'Rio Grande do Sul' },
-  { value: 'RO', label: 'Rondônia' },
-  { value: 'RR', label: 'Roraima' },
-  { value: 'SC', label: 'Santa Catarina' },
-  { value: 'SP', label: 'São Paulo' },
-  { value: 'SE', label: 'Sergipe' },
-  { value: 'TO', label: 'Tocantins' },
-];
-
-/** As escolhas que o campo oferece, já com a opção vazia na frente. */
-export const GENDER_SELECT_OPTIONS: readonly SelectOption[] = [
-  { value: '', label: SELECT_PLACEHOLDER },
-  ...GENDER_OPTIONS,
-];
-
-export const STATE_SELECT_OPTIONS: readonly SelectOption[] = [
-  { value: '', label: SELECT_PLACEHOLDER },
-  ...BRAZILIAN_STATES,
-];
