+++ FateConnect/Web/src/pages/Signup/styles.ts
@@ -89,39 +89,6 @@ export const FieldGrid = styled(Box)<PolymorphicProps>(({ theme }) => ({
   },
 }));
 
-/** Ocupa a linha inteira, em qualquer largura. */
-export const FullWidthCell = styled(Box)<PolymorphicProps>({
-  gridColumn: '1 / -1',
-  width: '100%',
-});
-
-/** Metade da linha no desktop: 3 de 6 colunas. */
-export const HalfWidthCell = styled(Box)<PolymorphicProps>({
-  width: '100%',
-
-  [desktopMedia]: { gridColumn: 'span 3' },
-});
-
-/** Um terço da linha no desktop: 2 de 6 colunas. */
-export const ThirdWidthCell = styled(Box)<PolymorphicProps>({
-  width: '100%',
-
-  [desktopMedia]: { gridColumn: 'span 2' },
-});
-
-/** Logradouro ocupa 4 de 6 colunas; o número fica ao lado. */
-export const StreetCell = styled(Box)<PolymorphicProps>({
-  width: '100%',
-
-  [desktopMedia]: { gridColumn: '1 / span 4' },
-});
-
-export const StreetNumberCell = styled(Box)<PolymorphicProps>({
-  width: '100%',
-
-  [desktopMedia]: { gridColumn: '5 / span 2' },
-});
-
 export const SubmitContainer = styled(Stack)<PolymorphicProps>({
   marginTop: spacing(md),
   flexDirection: 'column',
